rename files to match snake case
